requirements.txt
streamlit
requests
network
plotly
pandas
beautifulsoup4